+++ product.txt
@@ -1,6 +1,16 @@
-Мусорная корзина;0;Данный товар предназначен для выбрасывания в него мусора
+Мусорная корзина;1;Данный товар предназначен для выбрасывания в него мусора
 Машина;8;Да, мы имеем в своем списке товаров машину. Вас  что-то смущает?
 Утюг;8;Никто не будет в стороне
 Полотенце;8;Я ел собак
 Процессор;10;Не дорогой и древний
 Пиво;3;Вкусное
+Тарков;99;Душная игра для душнил
+КС;7;Шутер от первого лица
+Раки;15;Игроки в доту
+Креветки;45;Свежие, прям из моря
+Песок;56;Окенский, нормальный, тонкий
+Щебенка;65;Это ЩЕБЕЕЕЕНКА
+Кола;54;Газировка типичная
+Чипсики;21;Еда для котиков
+Сухарики;63;Праздник января
+Мармеладик;36;Мишки, если что
