+++ src/main/java/com/model/web_market/Person.java
@@ -1,4 +1,4 @@
-package com.control.web_market;
+package com.model.web_market;
 
 public class Person {
     private String name;
