+++ src/main/webapp/WEB-INF/web.xml
@@ -6,7 +6,7 @@
 
     <servlet>
         <servlet-name>MainPage</servlet-name>
-        <servlet-class>com.control.web_market.MainPage</servlet-class>
+        <servlet-class>com.model.web_market.MainPage</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>MainPage</servlet-name>
@@ -15,7 +15,7 @@
 
     <servlet>
         <servlet-name>Login</servlet-name>
-        <servlet-class>com.control.web_market.Login</servlet-class>
+        <servlet-class>com.model.web_market.Login</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>Login</servlet-name>
@@ -24,7 +24,7 @@
 
     <servlet>
         <servlet-name>Bag</servlet-name>
-        <servlet-class>com.control.web_market.Bag</servlet-class>
+        <servlet-class>com.model.web_market.Bag</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>Bag</servlet-name>
@@ -33,7 +33,7 @@
 
     <servlet>
         <servlet-name>Logout</servlet-name>
-        <servlet-class>com.control.web_market.Logout</servlet-class>
+        <servlet-class>com.model.web_market.Logout</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>Logout</servlet-name>
@@ -42,7 +42,7 @@
 
     <servlet>
         <servlet-name>AddBag</servlet-name>
-        <servlet-class>com.control.web_market.AddBag</servlet-class>
+        <servlet-class>com.model.web_market.AddBag</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>AddBag</servlet-name>
@@ -51,7 +51,7 @@
 
     <servlet>
         <servlet-name>AddProduct</servlet-name>
-        <servlet-class>com.control.web_market.AddProduct</servlet-class>
+        <servlet-class>com.model.web_market.AddProduct</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>AddProduct</servlet-name>
@@ -60,7 +60,7 @@
 
     <servlet>
         <servlet-name>removeBag</servlet-name>
-        <servlet-class>com.control.web_market.removeBag</servlet-class>
+        <servlet-class>com.model.web_market.removeBag</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>removeBag</servlet-name>
@@ -69,7 +69,7 @@
 
     <servlet>
         <servlet-name>removeProduct</servlet-name>
-        <servlet-class>com.control.web_market.removeProduct</servlet-class>
+        <servlet-class>com.model.web_market.removeProduct</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>removeProduct</servlet-name>
@@ -78,7 +78,7 @@
 
     <servlet>
         <servlet-name>ReProduct</servlet-name>
-        <servlet-class>com.control.web_market.ReProduct</servlet-class>
+        <servlet-class>com.model.web_market.ReProduct</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>ReProduct</servlet-name>
